refactor(rate-limit): reorganize structure
