+++ src/modules/_common/rate-limit/rateLimitTest.controller.ts
@@ -1,5 +1,5 @@
 import { Controller, Get, UseGuards } from "@nestjs/common";
-import { RateLimit } from "./rateLimit.decorator";
+import { RateLimit } from "./metadata/rateLimit";
 import { RateLimitGuard } from "./rateLimit.guard";
 
 @Controller("test-rate-limit")
