+++ src/modules/_common/rate-limit/types/RateLimitOptions.ts
@@ -1,0 +1,4 @@
+export interface RateLimitOptions {
+	ttl?: number; // in ms
+	limit?: number; // number of requests
+}
